concurrency updated formation added

+++ Huluwa2D/src/main/java/creatures/Moveable.java
@@ -1,5 +1,5 @@
 package creatures;
 
 public interface Moveable {
-    void move(double x, double y);
+    void move(int x, int y);
 }
